Developer bao

+++ src/main/java/com/Thienbao/uniclub/service/OrderService.java
@@ -1,16 +1,21 @@
 package com.Thienbao.uniclub.service;
 
+import com.Thienbao.uniclub.dto.OrderDetailDto;
+import com.Thienbao.uniclub.dto.OrderDto;
 import com.Thienbao.uniclub.exception.InsertOrderException;
 import com.Thienbao.uniclub.exception.NotFoundException;
-import com.Thienbao.uniclub.model.Cart;
-import com.Thienbao.uniclub.model.OrderDetail;
-import com.Thienbao.uniclub.model.Orders;
-import com.Thienbao.uniclub.model.User;
+import com.Thienbao.uniclub.exception.UpdateException;
+import com.Thienbao.uniclub.map.OrderDetailMapper;
+import com.Thienbao.uniclub.map.OrderMapper;
+import com.Thienbao.uniclub.model.*;
 import com.Thienbao.uniclub.model.key.OrderDetailID;
+import com.Thienbao.uniclub.model.key.ProductDetailID;
 import com.Thienbao.uniclub.payload.request.OrderRequest;
+import com.Thienbao.uniclub.payload.request.UpdateOrderRequest;
 import com.Thienbao.uniclub.repository.CartRepository;
 import com.Thienbao.uniclub.repository.OrderDetailRepository;
 import com.Thienbao.uniclub.repository.OrderRepository;
+import com.Thienbao.uniclub.repository.ProductDetailRepository;
 import com.Thienbao.uniclub.service.imp.OrderServiceImp;
 import com.Thienbao.uniclub.utils.JwtHelper;
 import jakarta.servlet.http.HttpServletRequest;
@@ -34,7 +39,16 @@ public class OrderService implements OrderServiceImp {
     private CartRepository cartRepository;
 
     @Autowired
+    private ProductDetailRepository productDetailRepository;
+
+    @Autowired
     private JwtHelper jwtHelper;
+
+    @Autowired
+    private OrderMapper orderMapper;
+
+    @Autowired
+    private OrderDetailMapper orderDetailMapper;
 
     @Transactional
     @Override
@@ -47,7 +61,7 @@ public class OrderService implements OrderServiceImp {
             order.setAddress(orderRequest.getAddress());
             order.setTotal(orderRequest.getTotal());
             order.setCreateDate(LocalDateTime.now());
-            order.setStatus("Pending");
+
             User user = new User();
             user.setId(idUser);
             order.setUser(user);
@@ -68,6 +82,20 @@ public class OrderService implements OrderServiceImp {
                 orderDetail.setQuantity(item.getQuantity());
                 orderDetailRepository.save(orderDetail);
 
+
+                // Subtract product quantity
+                ProductDetailID productDetailID = new ProductDetailID();
+                productDetailID.setIdProduct(item.getProduct().getId());
+                productDetailID.setIdColor(item.getColor().getId());
+                productDetailID.setIdSize(item.getSize().getId());
+
+                ProductDetail productDetail = productDetailRepository.findById(productDetailID).orElseThrow(()-> new NotFoundException("Not found Product detail with id"));
+
+                if(item.getQuantity() > productDetail.getQuantity()) throw new InsertOrderException("The number of products ordered cannot be greater than the number of products in stock");
+                int newQuantity = productDetail.getQuantity() - item.getQuantity();
+                productDetail.setQuantity(newQuantity);
+                productDetailRepository.save(productDetail);
+
                 // Delete Data in Cart Database.
                 cartRepository.delete(item);
             });
@@ -77,7 +105,37 @@ public class OrderService implements OrderServiceImp {
             throw new InsertOrderException("Error insert Order: " + ex.getMessage());
         }
 
+    }
+
+    @Override
+    public List<OrderDto> getAll() {
+        List<Orders> ordersList = orderRepository.findAll();
+        return ordersList.stream().map(orderMapper::convertToOrderDto).toList();
+    }
+
+    @Override
+    public List<OrderDetailDto> getAllOrderDetail(int idOrder) {
+        List<OrderDetail> orderDetails = orderDetailRepository.findAllByOrderId(idOrder);
+
+        if (orderDetails.isEmpty()) throw  new NotFoundException("Not found Orders with id order : " + idOrder);
+
+        return orderDetails.stream().map(orderDetailMapper::convertToOrderDetailDto).toList();
+
+    }
+
+    @Override
+    public boolean updateOrder(UpdateOrderRequest updateOrderRequest) {
+        Orders order = orderRepository.findById(updateOrderRequest.getIdOrder()).orElseThrow(()-> new NotFoundException("Not found order with order Id: " + updateOrderRequest.getIdOrder()));
+        try {
+            order.setStatus(updateOrderRequest.getStatus());
+            orderRepository.save(order);
+            return true;
+        }catch (Exception ex){
+            throw  new UpdateException("Error update order : " + ex.getMessage());
+        }
     };
+
+
 
 
 }
